+++ src/main/java/de/samply/form/pdf/FormPdfGeneratorFactory.java
@@ -11,7 +11,9 @@ import org.thymeleaf.spring6.SpringTemplateEngine;
 import org.thymeleaf.templatemode.TemplateMode;
 import org.thymeleaf.templateresolver.FileTemplateResolver;
 
+import java.io.File;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Path;
 
 @Component
 public class FormPdfGeneratorFactory {
@@ -22,22 +24,22 @@ public class FormPdfGeneratorFactory {
     public FormPdfGeneratorFactory(FormPdfConverter formPdfConverter,
                                    @Value(ProjectManagerConst.FORM_RESOURCES_DIRECTORY_SV) ExistingDirectory externalTemplateDirectory) {
         this.formPdfConverter = formPdfConverter;
-        this.templateEngine = createTemplateEngine(externalTemplateDirectory.path().toAbsolutePath().toString());
+        this.templateEngine = createTemplateEngine(externalTemplateDirectory.path().toAbsolutePath());
     }
 
     public PdfGenerator createPdfGenerator() {
         return new PdfGenerator(templateEngine, formPdfConverter);
     }
 
-    private TemplateEngine createTemplateEngine(String externalTemplateDirectory) {
+    private TemplateEngine createTemplateEngine(Path externalTemplateDirectory) {
         SpringTemplateEngine templateEngine = new SpringTemplateEngine();
         templateEngine.addTemplateResolver(createFileTemplateResolver(externalTemplateDirectory));
         return templateEngine;
     }
 
-    private FileTemplateResolver createFileTemplateResolver(String externalTemplateDirectory) {
+    private FileTemplateResolver createFileTemplateResolver(Path externalTemplateDirectory) {
         FileTemplateResolver resolver = new FileTemplateResolver();
-        resolver.setPrefix(externalTemplateDirectory);
+        resolver.setPrefix(externalTemplateDirectory + File.separator);
         resolver.setSuffix("." + FileExtension.HTML.value());
         resolver.setTemplateMode(TemplateMode.HTML);
         resolver.setCharacterEncoding(StandardCharsets.UTF_8.name());